fix: 유저상태관리 기능 오류 수정
